Moved functionality away from the overflow menu and Added a validations functionality
Moved delete all tasks and delete all done tasks from overflow menu to
the To-Do list
Added a validation functionality to the same two buttons to check if the
the array with task is empty and the array of done tasks is empty. This
is to prohibit the user to press on the buttons to delete something that
is already empty.